+++ ui/components/ItemDetail/CandidatureSpontanee/CandidatureSpontanee.js
@@ -57,7 +57,7 @@ const CandidatureSpontanee = (props) => {
 
   return (
     <div className="c-candidature" data-testid="CandidatureSpontanee">
-      <div className="c-detail-description-me col-12 col-md-5">
+      <div className="c-detail-description-me">
         <div className="c-detail-pelink my-3">
           {hasAlreadySubmittedCandidature({ applied, modal }) ? (
             <>
@@ -74,7 +74,7 @@ const CandidatureSpontanee = (props) => {
             <>
               <Button
                 onClick={toggle}
-                className={`btn btn-dark ml-1 gtmFormulaireCandidature gtm${capitalizeFirstLetter(kind)}`}
+                  className={`btn btn-blue ml-1 gtmFormulaireCandidature gtm${capitalizeFirstLetter(kind)}`}
                 aria-label="jenvoie-une-candidature-spontanee"
               >
                 J'envoie ma candidature{with_str(kind).amongst(["lbb", "lba"]) ? " spontanée" : ""}
@@ -119,6 +119,21 @@ const CandidatureSpontanee = (props) => {
             </>
           )}
         </div>
+        {
+          props?.item?.company?.mandataire ?
+          <>
+            <div className="c-mandataire-hint d-flex-center my-3">
+              <span className="c-mandataire-hintimg">
+                <img className="" src="/images/icons/small_info.svg" alt="point info"/>
+              </span>
+                <span className="c-mandataire-hinttext ml-2">
+                <span className="">Votre candidature sera envoyée au centre de formation en charge du recrutement pour le compte de l’entreprise.    </span>
+              </span>
+            </div>
+          </>
+          :
+          ""
+        }
       </div>
     </div>
   );