+++ ui/components/ItemDetail/Training.js
@@ -1,9 +1,10 @@
-import React, { useState } from "react";
+import React, { useState, useContext } from "react";
 import trainingIcon from "../../public/images/icons/book.svg";
 import { useSelector } from "react-redux";
 import { fetchAddresses } from "../../services/baseAdresse";
 import extendedSearchPin from "../../public/images/icons/jobPin.svg";
-import { useScopeContext } from "../../context/ScopeContext";
+//import { useScopeContext } from "../../context/ScopeContext";
+import { ScopeContext } from "../../context/ScopeContext";
 import { isCfaEntreprise } from "../../services/cfaEntreprise";
 import TagCfaDEntreprise from "./TagCfaDEntreprise";
 import { setSelectedMarker } from "../../utils/mapTools";
@@ -12,7 +13,7 @@ import { getSearchQueryParameters } from "../../utils/getSearchParameters";
 
 const Training = ({ training, handleSelectItem, showTextOnly, searchForJobsOnNewCenter }) => {
   const { formValues, itemParameters, selectedMapPopupItem } = useSelector((state) => state.trainings);
-  const scopeContext = useScopeContext();
+  const scopeContext = useContext(ScopeContext);
 
   const currentSearchRadius = formValues?.radius || 30;