+++ ui/context/ScopeContext.js
@@ -1,6 +1,6 @@
-import React from "react";
+import React, { createContext } from "react";
 
-const ScopeContext = React.createContext({
+export const ScopeContext = createContext({
   isTraining: true,
   isJob: true,
 });
@@ -9,5 +9,5 @@ export function ScopeContextProvider(props) {
   return <ScopeContext.Provider value={props.value}>{props.children}</ScopeContext.Provider>;
 }
 
-export const useScopeContext = () => React.useContext(ScopeContext);
-export default ScopeContext;
+/*export const useScopeContext = () => React.useContext(ScopeContext);*/
+//export default ScopeContext;